ingestion/requirements.txt
scrapy
python-dotenv